fix/tirando get da home e colocando no service
